Move selectors to their own module
