+++ qisbot/qis.py
@@ -1,4 +1,3 @@
-import enum
 import typing
 
 import requests
@@ -6,21 +5,11 @@ from lxml import html
 from lxml.etree import strip_tags
 
 from qisbot import scraper
+from qisbot.selectors import Selectors
 from qisbot.exceptions import NoSuchElementException
 from qisbot.exceptions import QisLoginFailedException
 from qisbot.exceptions import QisNotLoggedInException
 from qisbot.exceptions import UnexpectedStateException
-
-
-class Selectors(enum.Enum):
-    """Holds all non-trivial XPath expressions."""
-    LOGIN_ACTION_LINK = '//*[@id="wrapper"]/div[3]/a[2]'
-    EXAM_ADMINISTRATION_LINK = '//a[text() = "Prüfungsverwaltung" or text() = "Administration of exams"]'
-    EXAMS_EXTRACT_LINK = '//a[text() = "Notenspiegel" or text() = "Exams Extract"]'
-    SHOW_ACCOMPLISHMENTS_LINK = '//a[@title = "Leistungen anzeigen"]'
-    EXAMS_EXTRACT_EXAMS_TABLE = '//form/table[2]'
-    EXAMS_EXTRACT_EXAMS_CELLS = './/*[@class = "tabelle1_alignleft" or @class = "tabelle1_aligncenter ' \
-                                'or @class = "tabelle1_alignright"]'
 
 
 def requires_login(func):
